refactor: Test Fixture 클래스 전체 이동
