+++ matzip-app-external-api/src/test/java/com/woowacourse/matzip/application/RestaurantDemandServiceTest.java
@@ -1,8 +1,8 @@
 package com.woowacourse.matzip.application;
 
 import static com.woowacourse.auth.support.GithubResponseFixtures.ORI;
-import static com.woowacourse.matzip.support.RestaurantRequestFixtures.RESTAURANT_REQUEST_1;
-import static com.woowacourse.matzip.support.TestFixtureCreateUtil.createTestRestaurantDemand;
+import static com.woowacourse.matzip.RestaurantRequestFixtures.RESTAURANT_REQUEST_1;
+import static com.woowacourse.matzip.TestFixtureCreateUtil.createTestRestaurantDemand;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
