+++ matzip-core/src/testFixtures/java/com/woowacourse/matzip/ReviewFixtures.java
@@ -1,4 +1,4 @@
-package com.woowacourse.matzip.support;
+package com.woowacourse.matzip;
 
 import static com.woowacourse.auth.support.GithubResponseFixtures.HUNI;
 
@@ -6,7 +6,7 @@ import com.woowacourse.matzip.domain.member.Member;
 
 public enum ReviewFixtures {
 
-    REVIEW_1(HUNI.toMember(), "맛있네요 뽕쟁이", 4, "모듬족발(중)");
+    REVIEW_1(GithubResponseFixtures.HUNI.toMember(), "맛있네요 뽕쟁이", 4, "모듬족발(중)");
 
     private final Member member;
     private final String content;
